requirements.txt
python-dotenv==1.0.0
altair==4.2.2
streamlit==1.18.1
streamlit-chat==0.1.0
openai==0.27.7